Library Management system  (LMS)
The Library management system was developed to help small schools in the
purpose. It stores the books names in excel with their count. For each
borrow a seprate sheet with requied datails are generated for a user. A
seprate text file with necessary data about books present in library is
also incorporated.